testing mit trunk zusammengeführt

+++ linbo/init.sh
@@ -70,9 +70,18 @@ init_setup(){
  mount -t proc /proc /proc
  echo 0 >/proc/sys/kernel/printk
 
+ # parse kernel cmdline
  CMDLINE="$(cat /proc/cmdline)"
- echo "$CMDLINE" | grep -q debug && debug=yes
- echo "$CMDLINE" | grep -q useide && useide=yes
+ case "$CMDLINE" in *\ useide*) useide=yes;; esac
+ case "$CMDLINE" in *\ debug*) debug=yes;; esac
+ case "$CMDLINE" in *\ nonetwork*|*\ localmode*) localmode=yes;; esac
+ for i in $CMDLINE; do case "$i" in *=*) eval "$i";; esac; done
+
+ # get optionally given start.conf location
+ if [ -n "$conf" ]; then
+  confpart="$(echo $conf | awk -F\: '{ print $1 }')"
+  extraconf="$(echo $conf | awk -F\: '{ print $2 }')"
+ fi
 
  mount -t sysfs /sys /sys
  mount -n -o mode=0755 -t tmpfs tmpfs /dev
@@ -92,16 +101,6 @@ init_setup(){
  done
 }
 
-# findmodules dir
-# Returns module names for modprobe
-findmodules(){
- for m in `find "$@" -name \*.ko 2>/dev/null`; do
-  m="${m##*/}"
-  m="${m%%.ko}"
-  echo "$m"
- done
-}
-
 # trycopyfromdevice device filename
 trycopyfromdevice(){
  local RC=1
@@ -117,15 +116,64 @@ trycopyfromdevice(){
  fi
  return "$RC"
 }
- 
+
 # copyfromcache file - copies a file from cache to current dir
 copyfromcache(){
  local major="" minor="" blocks="" device="" relax=""
- [ -n "$cache" ] && trycopyfromdevice "$1" && return 0
+ [ -b "$cache" ] && trycopyfromdevice "$cache" "$1" && return 0
  while read major minor blocks device relax; do
    [ -n "$device" ] && trycopyfromdevice "$device" "$1" && return 0
  done < /proc/partitions
  return 1
+}
+
+# modify cache entry in start.conf
+modify_cache(){
+ [ -s "$1" ] || return 1
+ if grep -qi ^cache "$1"; then
+  sed -e "s|^[Cc][Aa][Cc][Hh][Ee].*|Cache = $cache|g" -i "$1"
+ else
+		sed -e "/^\[LINBO\]/a\
+Cache = $cache" -i "$1"
+ fi
+}
+
+# copytocache file - copies start.conf to local cache
+copytocache(){
+ local cachepart
+ if [ -b "$cache" ]; then
+  cachepart="$cache"
+ else
+  cachepart="$(grep -i ^cache /start.conf | tail -1 | awk -F\= '{ print $2 }' | awk '{ print $1 }')"
+ fi
+ case "$cachepart" in
+  /dev/*) # local cache
+   mount "$cachepart" /cache || return 1
+   cp -a /start.conf /cache
+   [ "$cachepart" = "$cache" ] && modify_cache /cache/start.conf
+   umount /cache || umount -l /cache
+   ;;
+  *)
+   echo "No local cache partition found!"
+   return 1
+   ;;
+ esac
+}
+
+# copy extra start.conf given on cmdline
+copyextra(){
+ [ -z "$confpart" ] && return 1
+ [ -z "$extraconf" ] && return 1
+ mkdir -p /extra
+ mount "$confpart" /extra || return 1
+ local RC=1
+ if [ -s "/extra$extraconf" ]; then
+  cp "/extra$extraconf" /start.conf ; RC="$?"
+  umount /extra || umount -l /extra
+ else
+  RC=1
+ fi
+ return "$RC"
 }
 
 # Try to read the first valid ip address from all up network interfaces
@@ -170,17 +218,17 @@ $(nslookup "$1" 2>/dev/null)
  return 1
 }
 
-# Get Gateway address as server.
+# Get server address.
 get_server(){
  local ip=""
  local a=""
  local b=""
- # Try servername from dhcp.log first.
+ # First try servername from dhcp.log:siaddr.
  if [ -f "/tmp/dhcp.log" ]; then
-  ip="`grep ^serverid /tmp/dhcp.log | tail -1 | cut -f2 -d"'"`"
+  ip="`grep ^siaddr /tmp/dhcp.log | tail -1 | cut -f2 -d"'"`"
   [ -n "$ip" ] && { echo "$ip"; return 0; }
  fi
- # Then guess from route
+ # Second guess from route.
  while read a b relax; do
   case "$a" in 0.0.0.0)
    ip="$b"
@@ -193,12 +241,22 @@ $(route -n)
  return 1
 }
 
+# get DownloadType from start.conf
+downloadtype(){
+ local RET=""
+ if [ -s /start.conf ]; then
+  RET="$(grep -i ^downloadtype /start.conf | tail -1 | awk -F= '{ print $2 }' | awk '{ print $1 }' | tr A-Z a-z)"
+  # get old option for compatibility issue
+  if [ -z "$RET" ]; then
+   RET="$(grep -i ^usemulticast /start.conf | tail -1 | awk -F= '{ print $2 }' | awk '{ print $1 }' | tr A-Z a-z)"
+   [ "$RET" = "yes" ] && RET="multicast"
+  fi
+ fi
+ echo "$RET"
+}
+
 network(){
- case "$(cat /proc/cmdline)" in *\ nonetwork|*\ localmode*)
-  touch /tmp/linbo-network.done
-  return 0
-  ;;
- esac
+ [ -n "$localmode" ] && { touch /tmp/linbo-network.done; return 0; }
  rm -f /tmp/linbo-network.done
  if [ -n "$ipaddr" ]; then
   [ -n "$netmask" ] && nm="netmask $netmask" || nm=""
@@ -221,19 +279,33 @@ network(){
  # Move away old start.conf and look for updates
  mv -f start.conf start.conf.dist
  if [ -n "$server" ]; then
+  echo "linbo_server='$server'" >> /tmp/dhcp.log
+  echo "mailhub=$server:25" > /etc/ssmtp/ssmtp.conf
   for i in "start.conf-$ipaddr" "start.conf"; do
    rsync -L "$server::linbo/$i" "start.conf" >/dev/null 2>&1 && break
   done
+  # also look for other needed files
+  case "$(downloadtype)" in
+   torrent) dlfile="torrent-client.conf" ;;
+   multicast) dlfile="multicast.list" ;;
+  esac
+  [ -n "$dlfile" ] && rsync -L "$server::linbo/$dlfile" "$dlfile" >/dev/null 2>&1
  fi
+ # copy start.conf optionally given on cmdline
+ copyextra && local extra=yes
  if [ ! -s start.conf ]; then
   # No new version / no network available, look for a cached copy.
-  while [ ! -e "/tmp/linbo-cache.done" ]; do # Wait unil harddisk is available
-   sleep 1
-  done
   copyfromcache start.conf
+ else
+  # flag for network connection
+  echo > /tmp/network.ok
+  # copy start.conf to cache if no extra start.conf was given on cmdline
+  [ -z "$extra" ] && copytocache
  fi
  # Still nothing new, revert to old version.
  [ -s start.conf ] || mv -f start.conf.dist start.conf
+ # modify cache in start.conf if cache was given and no extra start.conf was defined
+ [ -z "$extra" -a -b "$cache" ] && modify_cache /start.conf
  echo > /tmp/linbo-network.done 
 }
 
@@ -276,7 +348,17 @@ hwsetup(){
 }
 
 # Main
-echo "Hello, World."
+#clear
+echo
+echo 'Welcome to'
+echo ' _        _   __     _   ____      _____'
+echo '| |      | | |  \   | | |  _ \    / ___ \'
+echo '| |      | | |   \  | | | | | |  / /   \ \'
+echo '| |      | | | |\ \ | | | |/ /  | |     | |'
+echo '| |      | | | | \ \| | | |\ \  | |     | |'
+echo '| |____  | | | |  \   | | |_| |  \ \___/ /'
+echo '|______| |_| |_|   \__| |____/    \_____/'
+echo
 
 # Initial setup
 if [ -n "$debug" ]; then
